Moves the check_merged_patient into monitoring
Moves the check_merged_patient mgmt command into monitoring.
